bundle for text nodes
